requirements.txt
# ============================================
# ICEBOYS-bot Requirements (Stable, Production Ready)
# ============================================

# --- Core Telegram Bot Framework ---
python-telegram-bot==13.15 

# --- Network / API Handling ---
urllib3==1.26.16
requests==2.31.0
aiohttp==3.9.5
web3==6.15.1           # For Alchemy/On-Chain Payment Check

# --- Environment & Config ---
python-dotenv==1.0.1

# --- Logging & Utilities ---
colorama==0.4.6
psutil==6.0.0

# --- Database / Persistence ---
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9   # For PostgreSQL support

# --- Async Scheduling / Timing ---
apscheduler==3.6.3     # Fixed version for stability

# --- Optional AI & Parsing Modules ---
openai==1.13.3
pytz==2024.1
